requirements.txt
# UNIGEST - Dipendenze Python
# File: requirements.txt
# Descrizione: Elenco di tutte le librerie Python necessarie per il progetto

# Framework Django
Django==4.2.7

# Connettore MySQL per Django
mysqlclient==2.2.0

# Libreria per gestire variabili d'ambiente (file .env)
python-decouple==3.8

# Libreria per generare PDF (report)
reportlab==4.0.7

# Libreria per gestire file Excel (export/import)
openpyxl==3.1.2

# Libreria per gestire CSV in modo avanzato
pandas==2.1.3

# Libreria per widget avanzati nei form Django
django-widget-tweaks==1.5.0

# Libreria per filtrare dati nell'admin e nelle viste
django-filter==23.3

# Libreria per paginazione avanzata
django-bootstrap-pagination==1.7.2

# Pillow per gestione immagini (se necessario in futuro)
Pillow==10.1.0

# Django Debug Toolbar per sviluppo (opzionale ma consigliato)
django-debug-toolbar==4.2.0

# Libreria per backup automatici del database
django-dbbackup==4.0.2